Extract getResponseForPage
This was causing mocking and dependency cycle headaches so I have
extracted it to its own file
