+++ server/services/assessmentService.test.ts
@@ -11,15 +11,17 @@ import { getBody, updateAssessmentData } from '../form-pages/utils'
 import TasklistPage, { TasklistPageInterface } from '../form-pages/tasklistPage'
 import { DataServices, TaskListErrors } from '../@types/ui'
 import { ValidationError } from '../utils/errors'
-import { ApplicationOrAssessmentResponse, getResponseForPage } from '../utils/applications/utils'
+import { ApplicationOrAssessmentResponse } from '../utils/applications/utils'
 import { applicationAccepted } from '../utils/assessments/decisionUtils'
 import { getResponses } from '../utils/applications/getResponses'
+import { getResponseForPage } from '../utils/applications/getResponseForPage'
 
 jest.mock('../data/assessmentClient.ts')
 jest.mock('../data/personClient.ts')
 jest.mock('../form-pages/utils')
 jest.mock('../utils/applications/utils')
 jest.mock('../utils/applications/getResponses')
+jest.mock('../utils/applications/getResponseForPage')
 jest.mock('../utils/assessments/acceptanceData')
 jest.mock('../utils/assessments/decisionUtils')
